create cargo workspace and server entrypoint
This is the first of a series of commits of a major refactor. Almost a
rewrite. A workspace has been created that currently only references the
new `server` crate. All other modules outside that crate are legacy
code.

In addition to the new workspace the server entrypoint was established
in the `server` crate. It is a thread-per-connection TCP server. This is
different from the tokio based approach we used last time. I imagine we
will return to tokio fairly soon, but without proper benchmarks and a
working MVP it doesn't make sense to opt into that complexity until we
can prove that we need it.