FEAT: start in memory storage used during dev
This commit adds a new `storage` crate that defines a `Storage` struct.
Right now it is all in memory. This is because there are many many
things that have to be ironed out before it makes sense to write any of
the on-disk storage code. For instance the call-api of the handle,
mutability requirements, and lock/latch granularity. There's also a
chance that my current idea of the responsability boundary of such a
module is not fully correct yet.